Some more adjustments
Made route traversal speed based, with an adjustable modifier. Is not currently based on camera speed, but could be.
Angle changing is a bit off. Couldn't figure out the issue, so went with something that looked less weird. I added a TODO there
Fixed a route bug that would add an empty node if the route has an associated activity hash.
Route nodes are now affected by scaling.