Fix Route Spawning Command
CommandBuffers are out. Route specific actions are in.
Sadly can't get the entity to select them properly, since
the map may not be set up yet, so we can't reserve the entity
to select. We can, however, make a route specific action.

Also, wait for maps to exist before attempting to add a route.
We can do this by ending the ActivitySwapAction after a map is
loaded.