Add Ghost Mode
Ghost mode lets you display a ghost version of a structure in all maps.
This is useful for lining up an IRB by checking a target against other
maps seemlessly. The element is labeled as Ghost and Global for easy
tracking.

Currently this only works for some subset of things. For instance,
it works for static instance groups, but not static instances, at
least not once you swap maps.